Refactor animation of screen changes into a separate class
