+++ router/src/main/java/org/simple/clinic/router/screen/FullScreenKeyChanger.kt
@@ -2,13 +2,12 @@ package org.simple.clinic.router.screen
 
 import android.app.Activity
 import android.content.Context
-import androidx.annotation.ColorRes
-import androidx.annotation.IdRes
-import androidx.core.content.ContextCompat
-import androidx.interpolator.view.animation.FastOutSlowInInterpolator
 import android.view.View
 import android.view.ViewGroup
 import android.widget.FrameLayout
+import androidx.annotation.ColorRes
+import androidx.annotation.IdRes
+import androidx.core.content.ContextCompat
 import flow.Direction
 import flow.KeyChanger
 import flow.State
@@ -25,8 +24,9 @@ class FullScreenKeyChanger(
     private val activity: Activity,
     @IdRes private val screenLayoutContainerRes: Int,
     @ColorRes private val screenBackgroundRes: Int,
-    private val onKeyChange: (FullScreenKey?, FullScreenKey) -> Unit = { _, _ -> }
-) : BaseViewGroupKeyChanger<FullScreenKey>(), KeyChanger {
+    private val onKeyChange: (FullScreenKey?, FullScreenKey) -> Unit = { _, _ -> },
+    keyChangeAnimator: KeyChangeAnimator<FullScreenKey>
+) : BaseViewGroupKeyChanger<FullScreenKey>(keyChangeAnimator), KeyChanger {
 
   private val containerIds = mutableMapOf<FullScreenKey, Int>()
 
@@ -78,73 +78,5 @@ class FullScreenKeyChanger(
     val outgoingKey = outgoingState?.getKey<FullScreenKey>()
     val incomingKey = incomingState.getKey<FullScreenKey>()
     onKeyChange(outgoingKey, incomingKey)
-  }
-
-  override fun animate(
-      outgoingKey: FullScreenKey?,
-      outgoingView: View?,
-      incomingKey: FullScreenKey,
-      incomingView: View,
-      direction: flow.Direction,
-      onCompleteListener: () -> Unit
-  ) {
-    if (outgoingView == null || direction == flow.Direction.REPLACE) {
-      onCompleteListener.invoke()
-      return
-    }
-
-    val scaleChange = 0.05f
-    val duration = SCREEN_CHANGE_ANIMATION_DURATION
-    val interpolator = FastOutSlowInInterpolator()
-
-    if (direction == Direction.FORWARD || direction == Direction.REPLACE) {
-      outgoingView.animate()
-          .scaleX(1f - scaleChange)
-          .scaleY(1f - scaleChange)
-          .alpha(0f)
-          .setDuration(duration)
-          .setInterpolator(interpolator)
-          .start()
-
-    } else {
-      outgoingView.animate()
-          .scaleX(1f + scaleChange)
-          .scaleY(1f + scaleChange)
-          .alpha(0f)
-          .setDuration(duration)
-          .setInterpolator(interpolator)
-          .start()
-    }
-
-    if (direction == Direction.FORWARD || direction == Direction.REPLACE) {
-      incomingView.scaleX = 1f + scaleChange
-      incomingView.scaleY = 1f + scaleChange
-      incomingView.alpha = 0f
-      incomingView.animate()
-          .scaleX(1f)
-          .scaleY(1f)
-          .alpha(1f)
-          .setDuration(duration)
-          .setInterpolator(interpolator)
-          .withEndAction {
-            onCompleteListener.invoke()
-          }
-          .start()
-
-    } else {
-      incomingView.scaleX = 1f - scaleChange
-      incomingView.scaleY = 1f - scaleChange
-      incomingView.alpha = 0f
-      incomingView.animate()
-          .scaleX(1f)
-          .scaleY(1f)
-          .alpha(1f)
-          .setDuration(duration)
-          .setInterpolator(interpolator)
-          .withEndAction {
-            onCompleteListener.invoke()
-          }
-          .start()
-    }
   }
 }
